Pause works fully now. Changed from grabbing all of the scripts and disabling them to setting Time.timeScale to zero on pause and back on unpause.  Left the old means of pausing gameplay in the script, in case we need to go back to it after further testing. (Don't know if all of the scripts that move things work correctly after the timescale is set back to 1.)
Also I attached the WinCondition.cs to the player script so the ObjectiveScript can grab it.  In the process, I also assigned the Gold Globe graphic to the script for the winning item.  I don't know if that was correct for that script. The spot light on the win condition was left blank. (I didn't know if a specific light went there.)

Things to still do: Center pause menu box, and add in pause menu box designed for Occulus.  Is there a class for menu items to display correctly on the Occulus or do I have to set it up manually?