ALMOST THERE
Added in da shinies, but I wasn't able to implement the audio and the
last puzzle was giving me some issues, so I wasn't able to implement a
win condition. Beyond those two things the level is good to go. I will
be able to work on this again at 3:30, but I don't know when Tom is
checking the game out, so it might be too late for me to be able to fix
the last couple things. If someone has time earlier in the day and wants
to take a look and try to get those last pieces together (audio and win
condition), that would be awesome. Sorry I couldn't quite get those last
bits in.